% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accordion.R
\name{accordion}
\alias{accordion}
\title{accordion}
\usage{
accordion(id, items)
}
\arguments{
\item{id}{id of the accordion for html}

\item{items}{a taglist of accordion items}
}
\value{
shiny.tag
}
\description{
Create an accordion
}
\examples{
accordion(
  id = "myaccordion",
  items = tagList()
)

accordion(
  id = "myaccordion",
  items = tagList(
    accordion_item(
      accordion_id = "myaccordion",
      item_id = "accordition-item1",
      item_header = "Title 1",
      buttonid = "itemButton1",
      tags$p("This is a text"),
      tags$p("Inside an accordion item")
    ),
    accordion_item(
      accordion_id = "myaccordion",
      item_id = "accordition-item2",
      item_header = "Title 2",
      buttonid = "itemButton2",
      tags$p("This is a text"),
      tags$p("Inside an accordion item")
    )
  )
)
}
