% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accordion_item.R
\name{accordion_item}
\alias{accordion_item}
\title{accordion item}
\usage{
accordion_item(accordion_id, item_id, item_header, open, button_id, ...)
}
\arguments{
\item{accordion_id}{id of the accordion}

\item{item_id}{id of the accordion item}

\item{item_header}{header or title to display}

\item{open}{accordion item is open or collasped}

\item{button_id}{id of the button}

\item{...}{content to display in the accordion item body}
}
\value{
shiny.tag
}
\description{
Create an accordion item
}
\examples{
accordion_item(
  accordion_id = "accordion",
  item_id = "accordition-item",
  item_header = "Title",
  buttonid = "itemButton",
  tags$p("This is a text"),
  tags$p("Inside an accordion item")
)
}
