% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carousel.R
\name{carousel}
\alias{carousel}
\title{create a carousel}
\usage{
carousel(id, carousel_items)
}
\arguments{
\item{id}{id of the carousel for html}

\item{carousel_items}{a taglist of carousel_item()}
}
\value{
shiny.tag
}
\description{
create a carousel
}
\examples{
carousel(
  id = "carouselExample",
  tagList()
)

carousel(
  id = "carouselExampleIndicators",
  carousel_item = tagList(
    carousel_item(
      is_active = TRUE,
      img(src = "...")
    ),
    carousel_item(
      is_active = FALSE,
      img(src = "...")
    )
  )
)
}
