% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carousel_indicator.R
\name{carousel_indicator}
\alias{carousel_indicator}
\title{create a carousel indicator}
\usage{
carousel_indicator(id, carousel_item, pos)
}
\arguments{
\item{id}{id of the carousel}

\item{carousel_item}{carousel_item()}

\item{pos}{carousel item position in the taglist}
}
\value{
shiny.tag
}
\description{
create a carousel indicator
}
\examples{
carousel_indicator(
  id = "indic",
  carousel_item = carousel_item(
    is_active = TRUE,
    img(src = "...")
  ),
  pos = 1
)
}
