% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carousel_item.R
\name{carousel_item}
\alias{carousel_item}
\title{create a carousel item}
\usage{
carousel_item(is_active, ...)
}
\arguments{
\item{is_active}{is the carousel item active}

\item{...}{content to display in the carousel body}
}
\value{
shiny.tag
}
\description{
create a carousel item
}
\examples{
carousel_item(
  is_active = TRUE,
  img(src = "...")
)
}
