% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{tags}
\alias{tagList}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{htmltools}{\code{\link[htmltools]{tagList}}, \code{\link[htmltools:builder]{tags}}}
}}

